using Postgres DB

+++ src/main/java/dbschema/projects/upgrade_database_from_scripts_in_folder.sql
@@ -1,4 +1,4 @@
 --liquibase formatted sql
 
 --changeset upgrade_database_from_scripts_in_folder:1
-ALTER TABLE PERSON ADD AGE NUMBER(3);
+ALTER TABLE PERSON ADD AGE SMALLINT;
